This is our graphing web app
This is Version 0 of our web app. This web app takes a JSON file as a URL and plots it usingthe Plotly graphing library. This graph is sorted by dates and each coordinate contains a hash from the respective date of the commit. You can also use Plotly's integrated features to manipulate the graph (found in the top right corner of the graph).